added a demographic_report node
This node was formerly the "demographic" node. Now the demographic node has all the demographic core measures. This node has all the demographic properties used for the Explorer page.  In this updated version, a "multiracial" option was added to races and descriptions of how these properties compare to the actual core measures.